requirements.txt
streamlit
pandas
python-dotenv
groq
pymupdf
pillow
torch
torchvision
transformers
sentencepiece
accelerate
bitsandbytes
faiss-cpu
sentence-transformers